Merge environment vars with .env settings
The .env file is usually used by Heroku's Foreman to set environment
variables for local runs. We extend this idea and use this same file to
set up the environment in production as well, also on App Engine. This
allows removing common environment variables from app.yaml and
better conformance to DRY.